feat(workspace-cap): migrate timezone utilities (slice 9 of N preparatory)
ADR-152 section 3 F3.4 preparatory slice. Verbatim port of timezone.rs
(110 lines including 9 tests) from desktop-client/ironclaw/src/timezone.rs
into dasclaw_workspace_cap::timezone. ironclaw side becomes a 6-line
re-export shim, matching the slice 4-6 pattern. The two direct call sites
inside Workspace are switched to the cap path, removing Workspace's direct
dependency on crate::timezone. The 16 remaining call sites elsewhere in
ironclaw resolve through the shim and need no modification.

Closes #708
Refs #674
