+++ crates/dasclaw_workspace_cap/Cargo.toml
@@ -26,6 +26,8 @@ serde = { version = "1", features = ["derive"] }
 # Workspace memory modules (slice 3/N: privacy, document).
 regex = "1"
 chrono = { version = "0.4", features = ["serde"] }
+chrono-tz = "0.10"
+iana-time-zone = "0.1"
 uuid = { version = "1", features = ["v4", "serde"] }
 serde_json = "1"
