+++ crates/dasclaw_workspace_cap/src/lib.rs
@@ -55,6 +55,7 @@ pub mod policy;
 pub mod privacy;
 pub mod sanitization;
 pub mod search;
+pub mod timezone;
 
 pub mod error;
